Fix inspector type errors and improve tests
A couple of small bugs/inconsistencies in the inspector middleware.

1 - Setting the page size on the fly while using the inspector was not
working. Emacs was sending an integer to Cider, which was expecting a
string, causing a type conversion error (the tests were sending a
string, which is why it wasn't caught before). Also, the user can pass
in negative integers or a zero which would break the system, so added a
sanity check.

2 - Similarly, the push op expects a stringified integer in the :idx
slot. The elisp code was converting the supplied integer to a string
prior to sending it over the wire, so there was no error in the code,
but this has been changed so that the push op requires an integer and
the Cider Elisp code has been changed so that it will also send an
integer instead of unnecessarily converting it to a string. Please note
that if your Cider code is not as up to date as your Cider-nREPL code,
then the push function will not work.

Also added error handling try/catch blocks around all the messaging
operations. Added tests to exercise these.