Ensure CI deploy jobs run at the proper times
Fixing two errors by updating the conditional test in the deploy
stage in `.travis.yml`

1. While `make release` creates a new version tag from which the
release should build and deploy, it turns out the resulting CI job
does not run the deploy task, because for those jobs the branch name
is not "master".  With this commit we allow deploys from branches
where the name matches "v#.#.#".

2. Anyone running Travis CI builds on their own fork of this project
ended up with failed deploy jobs because the Clojars credentials could
not be parsed. With this commit we exclude deploy jobs from even
running if we're on a fork.