[#18] Fix doc "branch" name detection in CI
When "make docs" is invoked locally we can use `rev-parse` to
determine what branch we are on. However Travis CI checks out the
project with detached head, so we need to use $TRAVIS_BRANCH to figure
out where we are. As a result, we now have an if-then-else in the
Makefile recipe.