Add environment query functions
This was added to allow the sharing of query logic between various
different facets of cider.  Apropos, tests, and the namespace explorer
all expose different methods for operating on sets of vars from the
environment.

In particular, I want to do an apropos of only project namespaces, and
this isn't possible with current filtering rules.