Merge see-also in the info function
This patch takes the see-also code from the info middleware, merging the
:see-also key in here together with the rest of the info map.
It also does a very important thing, which is to convert all the strings in the
see-also.edn file to symbols. This is better for readability and performance
now that we are 100% sure (see info/normalize-params) that we are only passing
symbols downstream.