Improve approach to classloaders and classpaths
Rewrite classpath/classloader logic to work correctly with JDK9+, eliminate
hacks, and remove the `clojure.java.classpath` dependency.

Properly handle classpath entries as URLs to prevent platform-specific quirks
caused by treating them as Files.

Remove "fake" classpath references. It's either on the classpath or it isn't.

Don't rely on reflected access to classloader methods: update to `dynapath` 1.0.
This prevents access warnings in JDK9, which would become breakage in a future
JDK release.

When a call to `add-classpath!` is made, provide a modifiable classloader
if one is not already available rather than failing. This ensures equivalent
behavior between the REPL, where a modifiable classloader is in the hierarchy,
and running compiled code directly (`lein test`, `lein run`, `java ...`), where
only default classloaders are in the chain.

Closes #20
Closes #40