Don't try to use clojure.java.javadoc, just copy info
clojure.java.javadoc uses dynamic vars but stuffs them in a map so
its effectively no longer dynamic. Further, clojure 1.8 didn't have
javadoc url prefixes for jdk 11 so we just port over all the
information it was trying to do but in a better manner.

We could try to reuse the machinery in there but we would be updating
6 of 11 keys to whatever the proper api is so its easier to just copy
them over.