Allow for user added javadoc sources
In order to do so, need to use clojure.java.javadoc/*remove-javadocs*
but we update it with a proper version of the javadoc root. Some
issues are that clojure 1.8 has an entry for java 7 and for java 8 so
running clojure 1.8 on java 11 can't use javadocs without this
hack. Clojure 1.10.1 doesn't have a javadoc source for java 13 so this
ensures a consistent presence across all versions.