Only run unrecognized java version test on java 11
If we have an unrecognized java version, presumably it is higher than
11. It should have a notion of module name like 11 and 13. Running
this test on 8 breaks the test because there is no module name so that
part is missing from the slug. Seems the unrecognized versions will
most likely follow this path so we fallback to java 11 javadocs with a
module name rather than java 8 without them.