[Fix #86] Split java/resolve-symbol and shuffle clj-meta
A treaty could be written on the intricacies of clj-meta. Previously,
java/resolve-symbol was at the end of the chain and was not consulted, cause
conflicts in unqualified names, like max, would result in earlier returns (for
instance of clojure.core/max).

The problem is deeper and moving java/resolve-symbol up would not solve all the
problems, as it is doing two things: returning static member information AND
type information. Basically it broke records and macro refer resolution.

This patch splits that function in two: java/resolve-symbol only checks if
there exist Java members, and java/resolve-type only takes care of full classes
and records.

Interleaving the two with (m/resolve-var unqualified-sym) magically fulfills
all the use cases. There be dragons.