[java] Remove parser use of StandardDoclet
Remove the parser's use of `StandardDoclet` in favor of the reified
`getSupportedOptions` method returning the empty set. This is more accurate, as
the parser `Doclet` instance doesn't support any options. It also reduces
overall code path inside the JDK compiler classes, and therefore exposure to
future breakage like that documented in #100.

Closes #100