Further improve dynamic classpath modification robustness
When setting the context classloader, set this to newly created
DynamicClassLoader, then modify the highest visible modifiable ancestor.

The previous code, introduced in e3de281, set the context classloader to the
highest modifiable ancestor. This worked when launching via the `clojure`
executable, but caused visibility problems when launching with `lein`.