Fix info lookups from namespaces that don't yet exist
Currently when doing a `var-info` from a file before evaluating the `ns` form
always fails. The namespace object does not yet exist, and so none of the var
resolution strategies work.

However fully-qualified vars, namespace names, or built-ins (clojure.core
functions) could all be resolved, and it's frustrating that they aren't.

This adds a check to see if `(find-ns ns)` returns anything. If not then we are
initiating the lookup from a namespace that has not yet been defined, in which
case none of the lookup strategies will proceed, so we continue with doing the
lookup as if it had been initiated from `clojure.core`.