Fix problematic double var lookup in xref/fn-refs+fn-deps
`fn-deps` already calls `as-val` which will resolve a var if it is given one, so
`fn-refs` does not need to call `(map var-get)` beforehand. Not doing this
presents issues when a var contains a symbol. This currently would result in the
symbol being treated as a var name, causing exceptions or incorrect results.