* Analyze stacktrace in region or at point
This PR adds functionality to Orchard to support analyze stacktrace in region or at point commands in Cider.

** Problem

As a Clojure programmer I deal a lot with stacktraces. When evaluating code or dealing with the REPL, Cider has me covered here with its stacktrace inspector. Which is great!

But the stacktraces I'm looking at in the Cider stacktrace inspector are unfortunately only half of the stacktraces I am looking at in total. Many stacktraces I look at, I encounter in all sorts of Emacs buffers. Those stacktraces are printed in different formats like Java (.printStackTrace), Clojure (prn, #error tagged literal, clojure.stacktrace). Sometimes even buried within strings themselves or printed in a non-human friendly way. I would like to view those stacktraces pretty printed in the Cider inspector as well. With minimal effort, using Emacs commands like these:

- `M-x cider-stacktrace-analyze-in-region` Show the stacktrace in region in the Cider stacktrace inspector
- `M-x cider-stacktrace-analyze-at-point` Show the stacktrace at point in the Cider stacktrace inspector

** Solution

Let the client select the stacktrace to analyze (via `cider-stacktrace-analyze-in-region` for example), let the middleware parse and analyze it, and show the analysis in the client's stacktrace inspector.

** Approach

Add an optional "stacktrace" parameter to the Cider stacktrace middleware which the client can provide. If the "stacktrace" parameter is not blank, its value will be parsed and analyzed instead of the default `*e` session parameter. The middleware tries various strategies (different parsers for different formats, unwrapping an exception buried in a string, etc.) to produce a stacktrace analysis as a result.

** Limitations

- At the moment there's only one parser implemented in the middleware. The parser for the `#error` tagged literal, which Clojure uses to print stacktraces with `pr`. The idea is to make this plug-able by using a multi method for example. There could be more parsers for stacktraces produced by other printers like the Java `.printStackTrace` printer, the `clojure.stacktrace` printer, etc.

** Open questions

- What's the best approach to implement those parsers?
- Are the Clojure libraries out there that can parse already many formats?
- Can we depend on an external library for this if there is some benefit?
- Could we depend on Instaparse (supports Clojure since 1.5) in case we build our own parsers?