Catch NoClassDefFoundError
This exception can be thrown on a case indifferent filesystem, eg. on OS
X.

One way this happens is if there is a java class with a camel cased name
that otherwise matches the name of a clojure namespace.

For some reason, when this exception was thrown, the classes in the
class cache were also getting initialised.